+++ genomepy/providers/gencode.py
@@ -8,7 +8,7 @@ from genomepy.exceptions import GenomeDownloadError
 from genomepy.files import update_readme
 from genomepy.online import check_url, connect_ftp_link
 from genomepy.providers.base import BaseProvider, download_annotation
-from genomepy.providers.ncbi import download_assembly_report
+from genomepy.providers.ncbi import download_assembly_report, get_genome_size
 from genomepy.providers.ucsc import UcscProvider
 from genomepy.utils import get_genomes_dir, get_localname
 
@@ -48,13 +48,16 @@ class GencodeProvider(BaseProvider):
         """Can the provider be reached?"""
         return bool(check_url("ftp.ebi.ac.uk/pub/databases/gencode"))
 
-    def _genome_info_tuple(self, name):
+    def _genome_info_tuple(self, name, size=False):
         """tuple with assembly metadata"""
         accession = self.genomes[name]["assembly_accession"]
         taxid = self.genomes[name]["taxonomy_id"]
         annotations = True
         species = self.genomes[name]["species"]
         other = self.genomes[name]["other_info"]
+        if size:
+            length = get_genome_size(accession)
+            return name, accession, taxid, annotations, species, length, other
         return name, accession, taxid, annotations, species, other
 
     def _update_genomes(self):